move structs, docs
